Refactor service imports: move ResourceNotFoundException to utils package
