+++ server/src/main/java/com/kalyvianakis/estiator/io/service/IReservationService.java
@@ -2,7 +2,7 @@ package com.kalyvianakis.estiator.io.service;
 
 import java.util.List;
 
-import com.kalyvianakis.estiator.io.global.ResourceNotFoundException;
+import com.kalyvianakis.estiator.io.utils.ResourceNotFoundException;
 import com.kalyvianakis.estiator.io.model.Reservation;
 
 public interface IReservationService {
