+++ server/src/main/java/com/kalyvianakis/estiator/io/service/IUserService.java
@@ -1,6 +1,6 @@
 package com.kalyvianakis.estiator.io.service;
 
-import com.kalyvianakis.estiator.io.global.ResourceNotFoundException;
+import com.kalyvianakis.estiator.io.utils.ResourceNotFoundException;
 import com.kalyvianakis.estiator.io.model.User;
 
 import java.util.List;
